requirements.txt
openai
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0